LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/scroll-to-fragid/scroll-to-anchor-name-expected.txt


FAIL Fragment Navigation: scroll to anchor name is lower priority than equal id assert_equals: expected 400 but got 300
